requirements.txt
pymongo
pandas
mysql.connector
numpy
sshtunnel
typing
pymysql
python-dotenv
paramiko
pika
base